added Ui class
