+++ src/data/Yojdata.txt
@@ -1,3 +1,4 @@
 [T] | 0 | ok
 [T] | 1 | trial
 [T] | 0 | lol
+[T] | 0 | haha
